server/requirements.txt
fastapi
uvicorn
quantstats
panda
plotly
